Use ActiveRecord pattern for observers
Define own subclass of ActiveModel::Observer that takes care of hooking
into models and inheritance. As well as being a seemingly cleaner
separation of concerns, this allows us to handle single-columnfamily
inheritance properly in the context of observers.

Gets rid of Cequel::Model::Observing mixin -- functionality is defined
instead on Cequel::Model::Observer.