Revert back to instance-level use of Keyspace
Keyspace isn't meant to be globally configured; each Keyspace instance
can have a connection pool, which is lazily created if the configuration
specifies a pool size. Otherwise, it just lazy-initializes a single
instance variable, leaving it up to the user to make sure keyspaces are
being used in a thread-safe way.