Added reconnect and retry to Keyspace execute().
I did this because we were receiving RST packets during calls to execute()
after some amount of connection idle time. My assumption at this point is that
it is caused by a firewall or NetScaler in between our client using this gem
and the Cassandra server. Despite that though, this is a valid scenario that
needs handling because this is exactly what would happen if the Cassandra
server crashed and restarted. After the Cassandra restart call to execute()
would fail because the client still thinks the connection is up but the
Cassandra server knows nothing about it. Therefore, the Cassandra server would
send us a RST packet back, creating the same scenario as the first case.