Added logging of the rescued TransportException.
I added this logging so that when we handle a TransportException scenario it
doesn't just happen silently. This is important because this should NOT be
happening on a frequent basis by any means. This should happen once in a blue
moon and if it is happening frequently it is indicative of a problem with
either the Cassandra server or the configuration of the network components
between the client and the Cassandra server. Having this logging will allow us
to identify how frequent these issues are being handled.