Allow record sets to be scoped by multiple primary keys
E.g., `Post['blog1', 'blog2']`

If a partially-specified scope has multiple values for a key, it will execute
the scope with an IN query. If a fully-specified scope has multiple values for
a key, it will resolve to a wrapped array of unloaded records. When one of the
records has a property accessed, a single query will hydrate all of them.