Better interface for `::find`
As a general rule, passing varargs to `::find` maps those arguments onto
unscoped key columns; any single array argument scopes the find operation to
multiple values for that key.

In support of the principle of least astonishment, if there's only one unscoped
key column, varargs are treated as multiple values for that column. This is
undocumented.