Don't reset model updater/deleter if save results in an error
Since update operations are wrapped in a batch, actual error does not occur in
the code path of `#update`. This can result in a record being in a state where
it thinks it was successfully saved, but was not. Instead, we add the ability
to hook into a batch's completion, and then use that to only reset the model's
updaters when it is actually written to the database