Records don't automatically initiate logged batches
Previously, records would wrap operations in logged batches:

* Running the callback chain
* Issuing the statements needed to update a record's state (possibly
  both an `UPDATE` and a `DELETE`)

The latter is no longer needed at all, since records are updated purely
with an `UPDATE` statement in all cases.

The callback case does make some sense, but unfortunately if we're
running in a logged batch, we can't give immediate feedback on a
duplicate key. To keep things simple, and to get the much bigger
advantage of duplicate key detection, we remove the automatic batching
of callbacks.