Perform all row modifications in an UPDATE statement
In CQL3.1, anything we want to do to an existing row (besides deleting
the whole thing) can be done as part of an update statement, mostly by
using the `NULL` keyword, as well as the map subtraction syntax. So,
there's no need to use `DELETE` for these operations.

Thus, the `Deleter`'s job is reduced to purely that of deleting entire
rows. Because of this, it's no longer used by `Record` instances or
their collections to track discrete updates.

One pleasant side-effect is that updating a record will now always
happen in a single `UPDATE` statement.